@@ -3,13 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@supabase/supabase-js": "^1.35.7",
     "feather-icons": "^4.5.0",
-    "firebase": "^8.10.1",
-    "react": "^16.2.0",
+    "react": "^18.2.0",
     "react-cookie": "^2.1.2",
-    "react-dom": "^16.2.0",
-    "react-feather": "^1.0.8",
-    "react-scripts": "1.1.0"
+    "react-dom": "^18.2.0",
+    "react-feather": "^2.0.10",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -20,5 +21,17 @@
   "devDependencies": {
     "@babel/plugin-proposal-nullish-coalescing-operator": "^7.18.6",
     "@babel/plugin-proposal-optional-chaining": "^7.21.0"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
